# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education App Apis",
        "description": "Education App Apis",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courseChapters/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get All Chapters of Courses",
                "description": "Returns All Courses Chapters",
                "operationId": "courseChapters",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course Id to get Course Chapters",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/allCourses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get All Courses",
                "description": "Returns All Register Courses",
                "operationId": "course",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/single_course/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get Single Course",
                "description": "Returns All Register Courses",
                "operationId": "course",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course Id to get Single Course",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/SliderCourses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get All Courses For Slider",
                "description": "Returns All Courses For Slider",
                "operationId": "sliderCourses",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/PopularCourses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get All Popular Courses",
                "operationId": "PopularCourses",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/searchCourses/{keyword}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Search in All Courses against Keyword",
                "operationId": "searchCourses",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "description": "Get Courses By keyword",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/lessons/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get All Lesson of Courses",
                "description": "Returns All Register Course Content",
                "operationId": "lessons",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Chapter Id to get Chapter Lessons",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/currentUser": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get Current User",
                "description": "Returns Current User Data",
                "operationId": " ",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/UserProfileSave": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "User Profile Save",
                "description": "User Profile Save",
                "operationId": "User Profile Save",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "FirstName",
                                    "LastName",
                                    "gender",
                                    "pp_path",
                                    "city"
                                ],
                                "properties": {
                                    "FirstName": {
                                        "type": "string"
                                    },
                                    "LastName": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "InstituteName": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "pp_path": {
                                        "type": "file"
                                    },
                                    "city": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Education App Api",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}